真偽値による削除機能の実装(move uncorrectly)

+++ src/App.js
@@ -4,7 +4,7 @@ function App() {
   const [error, setError] = useState(null);
   const [isLoaded, setIsLoaded] = useState(false);
   const [datas, setDatas] = useState([]);
-  const [isFetched, setIsFetched] = useState(false)
+  const [isDeleted, setIsDeleted] = useState(false)
 
   useEffect(() => {
     fetch("https://api.manage.prona.com/client",{
@@ -14,8 +14,7 @@ function App() {
           'Access-Control-Allow-Origin': 'http://localhost:3000',
         },
         mode: 'cors',
-        
-      
+
     })
       .then(res => res.json())
       .then(
@@ -30,62 +29,55 @@ function App() {
         }
       )
   }, [])
-  
 
-  
   const clickDelete=(e)=>{
-   
-    let datas_id = datas.find(d => d.id === datas.id)
-    console.log(datas_id)
     
-      fetch("https://api.manage.prona.com/client/"+datas_id ,{
-      method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-          'Access-Control-Allow-Origin': 'http://localhost:3000', 
-        },
-        
-        mode: 'cors',
-        body: datas //DELETEメソッド使用時にjsonにパースするとエラーが起きるためそのまま渡しました。
-       })
-       .then(response=> response.text())
-       .then((result)=>{
-         setIsFetched(true);
-         console.log(result);
-        //fetchが成功=isFetchedがtrueの時なので、以下の配列削除を実行
-        const deleteDatas = datas.slice();
-        deleteDatas.splice(e,1);
-        setDatas(deleteDatas);
-        e.preventDefault();
-
-        if(!isFetched){
-          return <div>削除中...</div>
-        }else{
-          return <div>削除しました。</div>
-        }
-       })
+        if(!isDeleted){
+          fetch("https://api.manage.prona.com/client/"+datas[0].id ,{
+             method: 'DELETE',
+              headers: {
+                'Content-Type': 'application/json',
+                'Access-Control-Allow-Origin': 'http://localhost:3000', 
+              },
+              mode: 'cors',
+          })
+          .then(response=> response.text())
+          .then((result)=>{
+            setIsDeleted(true);
+            console.log(result);
+          })
+          
+          }else{
+            const deleteDatas = datas.slice();
+            deleteDatas.splice(e,1);
+            setDatas(deleteDatas);
+            setIsDeleted(false)
+            e.preventDefault();
+          }
+       
   }
-
       if (error) {
         return <div>Error: {error.message}</div>;
       } else if (!isLoaded) {
         
         return <div>Loading...</div>;
       } else{
+        
       return <div>
         {datas.map(item => (
           <form key={item.id}>
             <div>{item.name}</div>
             <br />
-            <div>{item.email}</div>
-            <br />
-            <div>{item.hash}</div>
-            <br />
             <div>{item.id}</div>
             <br />
             <input onClick={clickDelete} type="button" value="削除" />
+            <br />
           </form>
-          ))}</div>
+          ))}
+          {isDeleted?  <div>削除中...</div> 
+          : 
+          <div>削除しました。</div>}
+          </div>
     }
 }
 export default App;
